decoupled the download route controller
